requirements.txt
Django>=4.2
requests>=2.28
gunicorn>=20.1
python-dotenv>=1.0
google-api-python-client>=2.70.0
google-auth>=2.20.0
google-auth-httplib2>=0.1.0
requests-oauthlib>=1.3.1

# Optional MongoDB support. To enable connecting Django to MongoDB via djongo,
# install djongo and pymongo separately in your environment. We don't force
# djongo in requirements here to avoid installation errors on environments
# that don't need it.
# djongo==1.3.7
# pymongo[srv]>=4.6.0
